chore: use ops network object instead of mocking get_binding
